mesh_core/mesh_node/requirements.txt
PyYAML==6.0.1
cryptography==43.0.1
paho-mqtt==2.1.0
python-lora==0.6.0
python-pbft==0.2.1
numpy==1.26.4
protobuf==5.26.1
pyserial==3.5
boto3==1.35.10
cbor2==5.6.2
redis==5.0.8
pyotp==2.9.0
argon2-cffi==23.1.0
scipy==1.13.0
python-dotenv==1.0.1